+++ core/exec/test/src/mill/exec/EvaluatorPathsTests.scala
@@ -1,4 +1,4 @@
-package mill.eval
+package mill.exec
 
 import mill.exec.ExecutionPaths
 import utest.*
