+++ core/exec/test/src/mill/exec/FailureTests.scala
@@ -1,4 +1,4 @@
-package mill.eval
+package mill.exec
 
 import mill.Task
 import mill.testkit.UnitTester
