+++ core/exec/test/src/mill/exec/JavaCompileJarTests.scala
@@ -1,4 +1,4 @@
-package mill.eval
+package mill.exec
 
 import mill.util.Jvm
 import mill.api.Ctx.Dest
