+++ core/exec/test/src/mill/exec/ModuleTests.scala
@@ -1,4 +1,4 @@
-package mill.eval
+package mill.exec
 
 import mill.testkit.UnitTester
 import mill.testkit.UnitTester.Result
