+++ core/exec/test/src/mill/exec/TaskTests.scala
@@ -1,4 +1,4 @@
-package mill.eval
+package mill.exec
 
 import utest.*
 import mill.Task
